tests/control-software-validator/requirements.txt
pytest==8.3.5
pydantic==2.10.6
datamodel-code-generator==0.28.4
pyzmq==26.3.0
pytest-asyncio==0.25.0
